refactor: change dependency direction (#87)
* chore: add auto generated qclass to .gitignore

* fix: failed test by current time

* refactor: apply dip for package dependency

* chore: update interface name

* test: modify test data

* fix: test code
