+++ api-server/src/main/java/nexters/payout/apiserver/stock/application/StockDividendQueryService.java
@@ -1,0 +1,7 @@
+package nexters.payout.apiserver.stock.application;
+
+import nexters.payout.apiserver.stock.application.dto.response.StockDetailResponse;
+
+public interface StockDividendQueryService {
+    StockDetailResponse getStockByTicker(final String ticker);
+}
